crypto/requirements.txt
django==5.1.6
requests==2.31.0
web3==6.20.0
python-decouple==3.8
whitenoise==6.7.0
gunicorn==23.0.0